+++ CHANGELOG.md
@@ -28,6 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - PAGES: 一部のページの文章が変更されました
 - PAGES: ストレージ管理の警告メッセージが設定からページ内に移動しました
 - PAGES: 「現在のタブから追加可能なシリーズ」が新視聴ページからも検出できるようになりました
+- PAGES: React Icons の Material Icons から Tabler Icons に移行しました
+- PAGES: ヘッダーのリンクが一部アイコンに変更されました
 - DMS: `NordNight` のパレット割当が変更されました
 - DMS: 動画以外のダークモードサポートが個別のstylファイルに分解されました